feat: Adicionar crud de usuários e auth

+++ src/app/services/user/DeleteUserService.service.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from "tsyringe";
 import type { IUserRepository } from "../../repositories/user/IUserRepository.js";
 import { deleteUserSchema } from "../../validations/user/deleteUser.schema.js";
+import { AppError } from "../../errors/appError.js";
 
 @injectable()
 export default class DeleteUserService {
@@ -11,6 +12,13 @@ export default class DeleteUserService {
 
   async execute(params: { id: string }): Promise<void> {
     const { id } = deleteUserSchema.parse({ id: params.id });
+
+    const user = await this.userRepository.findById(id);
+    
+    if (!user) {
+      throw new AppError("Usuário não encontrado", 404);
+    }
+
     await this.userRepository.delete(id);
     return;
   }
